Install missing components from known config for all stores, not just steam
